# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Parcel Intelligence API
  version: 0.1.0
paths:
...
  /api/v1/parcels:
    get:
      tags:
      - parcels
      summary: Search Parcels
      description: Search and filter cadastral land parcels across administrative
        hierarchies.
      operationId: search_parcels_api_v1_parcels_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search by ULPIN, Survey Number, Khasra Number, or Patta Number
          title: Q
        description: Search by ULPIN, Survey Number, Khasra Number, or Patta Number
      - name: state_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Filter by State ID
          title: State Id
        description: Filter by State ID
      - name: district_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Filter by District ID
          title: District Id
        description: Filter by District ID
      - name: village_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Filter by Village ID
          title: Village Id
        description: Filter by Village ID
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParcelSummary'
                title: Response Search Parcels Api V1 Parcels Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}:
    get:
      tags:
      - parcels
      summary: Get Parcel Detail
      description: Retrieve comprehensive cadastral parcel details including GeoJSON
        polygon boundary.
      operationId: get_parcel_detail_api_v1_parcels__ulpin__get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParcelDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}/ror:
    get:
      tags:
      - parcels
      summary: Get Parcel Ror
      description: Retrieve Record of Rights (RoR / Khatauni) ownership records for
        a parcel.
      operationId: get_parcel_ror_api_v1_parcels__ulpin__ror_get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoRItem'
                title: Response Get Parcel Ror Api V1 Parcels  Ulpin  Ror Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}/registrations:
    get:
      tags:
      - parcels
      summary: Get Parcel Registrations
      description: Retrieve Sub-Registrar registered deed transactions for a parcel.
      operationId: get_parcel_registrations_api_v1_parcels__ulpin__registrations_get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegistrationItem'
                title: Response Get Parcel Registrations Api V1 Parcels  Ulpin  Registrations
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}/encumbrances:
    get:
      tags:
      - parcels
      summary: Get Parcel Encumbrances
      description: Retrieve bank mortgages, judicial liens, or litigation claims on
        a parcel.
      operationId: get_parcel_encumbrances_api_v1_parcels__ulpin__encumbrances_get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EncumbranceItem'
                title: Response Get Parcel Encumbrances Api V1 Parcels  Ulpin  Encumbrances
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}/mutations:
    get:
      tags:
      - parcels
      summary: Get Parcel Mutations
      description: Retrieve land title mutation and transfer history for a parcel.
      operationId: get_parcel_mutations_api_v1_parcels__ulpin__mutations_get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MutationItem'
                title: Response Get Parcel Mutations Api V1 Parcels  Ulpin  Mutations
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}/building-permissions:
    get:
      tags:
      - parcels
      summary: Get Parcel Building Permissions
      description: Retrieve town planning sanctioned building plans and permissions.
      operationId: get_parcel_building_permissions_api_v1_parcels__ulpin__building_permissions_get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuildingPermissionItem'
                title: Response Get Parcel Building Permissions Api V1 Parcels  Ulpin  Building
                  Permissions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}/tax:
    get:
      tags:
      - parcels
      summary: Get Parcel Property Tax
      description: Retrieve municipal property tax assessments and payment statuses.
      operationId: get_parcel_property_tax_api_v1_parcels__ulpin__tax_get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PropertyTaxItem'
                title: Response Get Parcel Property Tax Api V1 Parcels  Ulpin  Tax
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}/history:
    get:
      tags:
      - parcels
      summary: Get Parcel Combined History
      description: Retrieve an integrated chronological timeline merging registrations,
        mutations, and alerts.
      operationId: get_parcel_combined_history_api_v1_parcels__ulpin__history_get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParcelHistoryItem'
                title: Response Get Parcel Combined History Api V1 Parcels  Ulpin  History
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}/zoning-check:
    post:
      tags:
      - parcels
      summary: Check Parcel Zoning
      description: Perform spatial intersection check between parcel geometry and
        master plan zoning.
      operationId: check_parcel_zoning_api_v1_parcels__ulpin__zoning_check_post
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoningCheckResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/parcels/{ulpin}/intelligence:
    get:
      tags:
      - parcels
      summary: Get Parcel Intelligence
      description: Calculate and retrieve statutory dispute risk, title trust score,
        and explainable deductions.
      operationId: get_parcel_intelligence_api_v1_parcels__ulpin__intelligence_get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisputeIntelligenceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/public/parcels/{ulpin}:
    get:
      tags:
      - developer
      summary: Get Public Parcel Data
      description: Public, sanitized parcel lookup endpoint (excludes citizen PII
        and sensitive legal encumbrances).
      operationId: get_public_parcel_data_api_v1_public_parcels__ulpin__get
      parameters:
      - name: ulpin
        in: path
        required: true
        schema:
          type: string
          title: Ulpin
      - name: x-api-key
        in: header
        required: true
        schema:
          type: string
          description: Developer Sandbox API Key
          title: X-Api-Key
        description: Developer Sandbox API Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicParcelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EncumbranceItem:
      properties:
        id:
          type: integer
          title: Id
        type:
          type: string
          title: Type
        holder_name:
          type: string
          title: Holder Name
        amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Amount
        start_date:
          type: string
          format: date
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        status:
          type: string
          title: Status
      type: object
      required:
      - id
      - type
      - holder_name
      - start_date
      - status
      title: EncumbranceItem
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MutationItem:
      properties:
        id:
          type: integer
          title: Id
        mutation_type:
          type: string
          title: Mutation Type
        previous_owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Previous Owner
        new_owner:
          anyOf:
          - type: string
          - type: 'null'
          title: New Owner
        applied_date:
          type: string
          format: date
          title: Applied Date
        approved_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Approved Date
        status:
          type: string
          title: Status
        remarks:
          anyOf:
          - type: string
          - type: 'null'
          title: Remarks
      type: object
      required:
      - id
      - mutation_type
      - applied_date
      - status
      title: MutationItem
    ParcelDetail:
      properties:
        id:
          type: integer
          title: Id
        ulpin:
          type: string
          title: Ulpin
        survey_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Survey Number
        khasra_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Khasra Number
        gata_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Gata Number
        patta_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Patta Number
        area_recorded_sqm:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Recorded Sqm
        area_gis_sqm:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Gis Sqm
        land_use_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Land Use Type
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
        district:
          anyOf:
          - type: string
          - type: 'null'
          title: District
        village:
          anyOf:
          - type: string
          - type: 'null'
          title: Village
        geometry:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Geometry
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - ulpin
      title: ParcelDetail
...
    ParcelSummary:
      properties:
        id:
          type: integer
          title: Id
        ulpin:
          type: string
          title: Ulpin
        survey_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Survey Number
        khasra_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Khasra Number
        gata_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Gata Number
        patta_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Patta Number
        area_recorded_sqm:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Recorded Sqm
        area_gis_sqm:
          anyOf:
          - type: number
          - type: 'null'
          title: Area Gis Sqm
        land_use_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Land Use Type
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
        district:
          anyOf:
          - type: string
          - type: 'null'
          title: District
        village:
          anyOf:
          - type: string
          - type: 'null'
          title: Village
      type: object
      required:
      - id
      - ulpin
      title: ParcelSummary
    PropertyTaxItem:
      properties:
        id:
          type: integer
          title: Id
        assessment_year:
          type: integer
          title: Assessment Year
        assessed_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Assessed Value
        tax_amount:
          type: number
          title: Tax Amount
        paid_status:
          type: string
          title: Paid Status
        ulb_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Ulb Id
      type: object
      required:
      - id
      - assessment_year
      - tax_amount
      - paid_status
      title: PropertyTaxItem
...
    RegistrationItem:
      properties:
        id:
          type: integer
          title: Id
        deed_type:
          type: string
          title: Deed Type
        deed_number:
          type: string
          title: Deed Number
        registration_date:
          type: string
          format: date
          title: Registration Date
        sub_registrar_office:
          anyOf:
          - type: string
          - type: 'null'
          title: Sub Registrar Office
        consideration_amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Consideration Amount
        ngdrs_ref_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Ngdrs Ref Id
        document_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Hash
      type: object
      required:
      - id
      - deed_type
      - deed_number
      - registration_date
      title: RegistrationItem
    RoRItem:
      properties:
        id:
          type: integer
          title: Id
        owner_name:
          type: string
          title: Owner Name
        ownership_type:
          type: string
          title: Ownership Type
        share_percentage:
          type: number
          title: Share Percentage
        tenure_type:
          type: string
          title: Tenure Type
        khatauni_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Khatauni Number
        source_document_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Document Ref
        valid_from:
          type: string
          format: date
          title: Valid From
        valid_to:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Valid To
        status:
          type: string
          title: Status
      type: object
      required:
      - id
      - owner_name
      - ownership_type
      - share_percentage
      - tenure_type
      - valid_from
      - status
      title: RoRItem
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer